gateways/requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
requests==2.32.3
urllib3==2.2.2
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
aiokafka==0.10.0
flasgger==0.9.7.1
pydantic==2.8.2
python-dateutil==2.9.0